* egil/darwin-fixes:
  erl_interface: Avoid redefinition of ALIGNBYTES
  erts: Fix '#ifdef' for unused static functions
  erts: Remove forgotten vxworks includes
  hipe: Cast offsetof to int in hipe_arch_print_pcb

+++ erts/emulator/hipe/hipe_x86.c
@@ -265,7 +265,7 @@ void *hipe_make_native_stub(void *beamAddress, unsigned int beamArity)
 void hipe_arch_print_pcb(struct hipe_process_state *p)
 {
 #define U(n,x) \
-    printf(" % 4d | %s | 0x%08x |            |\r\n", offsetof(struct hipe_process_state,x), n, (unsigned)p->x)
+    printf(" % 4d | %s | 0x%08x |            |\r\n", (int)offsetof(struct hipe_process_state,x), n, (unsigned)p->x)
     U("ncsp       ", ncsp);
     U("narity     ", narity);
 #undef U
